refactor(preflight): split plan contract validation into focused modules
